[ISSUE-9] - migration string resources: improvements
