+++ core-shared-ui/build.gradle.kts
@@ -24,3 +24,6 @@ kotlin {
     }
 }
 
+compose.resources {
+    publicResClass = true
+}
