--- core-shared-ui/src/commonMain/kotlin/com/codandotv/streamplayerapp/core_shared_ui/resources/StringCoreSharedUi.kt
@@ -1,23 +1,0 @@
-package com.codandotv.streamplayerapp.core_shared_ui.resources
-
-import streamplayerapp_kmp.core_shared_ui.generated.resources.Res
-import streamplayerapp_kmp.core_shared_ui.generated.resources.app_name
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_back
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_cast
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_close
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_mic
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_netflix
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_profile
-import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_search
-
-object StringCoreSharedUi {
-    val AppName = Res.string.app_name
-    val IconNetflix = Res.string.icon_netflix
-    val IconProfile = Res.string.icon_profile
-    val IconSearch = Res.string.icon_search
-    val IconClose = Res.string.icon_close
-    val IconMic = Res.string.icon_mic
-    val IconBack = Res.string.icon_back
-    val IconCast = Res.string.icon_cast
-
-}
