+++ core-shared-ui/src/commonMain/kotlin/com/codandotv/streamplayerapp/core_shared_ui/widget/SearchIcon.kt
@@ -1,0 +1,19 @@
+package com.codandotv.streamplayerapp.core_shared_ui.widget
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import org.jetbrains.compose.resources.stringResource
+import streamplayerapp_kmp.core_shared_ui.generated.resources.Res
+import streamplayerapp_kmp.core_shared_ui.generated.resources.icon_search
+
+@Composable
+fun SearchIcon(action: () -> Unit = {}) {
+    DefaultIcon(
+        searchIcon = Icons.Filled.Search,
+        contentDescription = stringResource(Res.string.icon_search),
+        onIconClickAction = action,
+        iconColor = Color.Gray
+    )
+}
