+++ feature-search/src/commonMain/kotlin/com/codandotv/streamplayerapp/feature_search/presentation/widgets/SearchStreams.kt
@@ -35,9 +35,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.codandotv.streamplayerapp.core.shared.ui.R
 import com.codandotv.streamplayerapp.core_shared_ui.resources.Colors
-import com.codandotv.streamplayerapp.core_shared_ui.resources.StringCoreSharedUi
+import com.codandotv.streamplayerapp.core_shared_ui.widget.CloseButton
+import com.codandotv.streamplayerapp.core_shared_ui.widget.MicButton
+import com.codandotv.streamplayerapp.core_shared_ui.widget.SearchIcon
 import org.jetbrains.compose.resources.stringResource
+import streamplayerapp_kmp.core_shared_ui.generated.resources.*
 import streamplayerapp_kmp.feature_search.generated.resources.Res
+import streamplayerapp_kmp.core_shared_ui.generated.resources.Res as SharedRes
 import streamplayerapp_kmp.feature_search.generated.resources.search_list_main_search
 
 @Suppress("LongParameterList")
@@ -80,7 +84,7 @@ private fun StreamPlayerTopBar(
         ) {
             Icon(
                 imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                contentDescription = stringResource(StringCoreSharedUi.IconBack),
+                contentDescription = stringResource(SharedRes.string.icon_back),
                 tint = Color.White,
             )
         }
@@ -92,7 +96,7 @@ private fun StreamPlayerTopBar(
         ) {
             Icon(
                 imageVector = Icons.Default.Check,
-                contentDescription = stringResource(StringCoreSharedUi.IconCast),
+                contentDescription = stringResource(SharedRes.string.icon_cast),
                 tint = Color.White,
             )
         }
@@ -106,7 +110,7 @@ private fun StreamPlayerTopBar(
                     .height(24.dp)
                     .clip(RoundedCornerShape(4.dp)),
                 painter = painterResource(R.drawable.perfil_fake),
-                contentDescription = stringResource(StringCoreSharedUi.IconProfile),
+                contentDescription = stringResource(SharedRes.string.icon_profile),
                 tint = Color.Unspecified,
             )
         }
@@ -168,55 +172,6 @@ fun SearchTopBar(
     }
 }
 
-@Composable
-fun DefaultIcon(
-    modifier: Modifier = Modifier,
-    searchIcon: ImageVector = Icons.Default.Search,
-    iconColor: Color = Color.White,
-    contentDescription: String = "",
-    onIconClickAction: () -> Unit = {}
-) {
-    IconButton(
-        modifier = modifier,
-        onClick = onIconClickAction
-    ) {
-        Icon(
-            imageVector = searchIcon,
-            contentDescription = contentDescription,
-            tint = iconColor
-        )
-    }
-}
-
-@Composable
-fun SearchIcon(action: () -> Unit = {}) {
-    DefaultIcon(
-        searchIcon = Icons.Filled.Search,
-        contentDescription = stringResource(StringCoreSharedUi.IconSearch),
-        onIconClickAction = action,
-        iconColor = Color.Gray
-    )
-}
-
-@Composable
-fun CloseButton(action: () -> Unit = {}) {
-    DefaultIcon(
-        searchIcon = Icons.Default.Close,
-        contentDescription = stringResource(StringCoreSharedUi.IconClose),
-        onIconClickAction = action,
-        iconColor = Color.Gray
-    )
-}
-
-@Composable
-private fun MicButton(action: () -> Unit = {}) {
-    DefaultIcon(
-        searchIcon = Icons.Default.Check,
-        contentDescription = stringResource(StringCoreSharedUi.IconMic),
-        onIconClickAction = action,
-        iconColor = Color.Gray
-    )
-}
 
 @Composable
 @Preview
